requirements.txt
asgiref==3.11.0
Django==5.2.9
django-decouple==2.1
djangorestframework==3.16.1
dotenv==0.9.9
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.5
